Removal Neural Libraries and Stabilized Trader Game
Still have not tested the trader games stabilizing of the ship
controller class, as well as a new implementation seperating the
SolarBody class from the orbit class

+++ Trader/Assets/_Game/Scripts/Helpers/Items.cs
@@ -4,8 +4,8 @@ using System.Collections;
 public class Items {
     public string name;
     public ItemTypes itemType;
-    public int pendingAmount;
-    public int amount;
+    public float pendingAmount;
+    public float amount;
     public Color color;
     public string coloredName;
